cleanup: move .getNumber to a different util file
